fix(R-02,R-04,R-05,R-08,R-09): medium-priority architecture violations
R-02: extract GetCreditsByBroker/GetCreditsByPortfolio from NavigationService into
a dedicated CreditQueryService implementing ICreditQueryService; update DI registration
to bind ICreditQueryService to CreditQueryService; update NavigationServiceTests to use
CreditQueryService for credit-related assertions.

R-04: make Credit.EnsureId() internal (matching Transaction.EnsureId which was already
internal); add [InternalsVisibleTo("Financial.Domain.Tests")] so the domain test can
still call it directly.

R-05: remove static HttpClient field from AssetTypeLookup; inject HttpClient via
constructor; register IAssetTypeLookup in InfrastructureServiceCollectionExtensions
using a singleton factory.

R-08: move DefaultDividendWatchlist out of MainWindow static field into WatchlistOptions;
extract Option nested class to WatchlistItem in its own file; MainWindow now reads from
WatchlistOptions.DefaultDividendWatchlist.

R-09: extract GetPaddedShortDatePattern, PadDayMonthTokens, HandleQuoteChar, HandleFormatToken
from MainWindow to a dedicated DateFormatHelper static class; remove the duplicated private
methods from MainWindow.
