+++ Financial.App/Options/WatchlistItem.cs
@@ -1,0 +1,7 @@
+namespace Financial.Presentation.App.Options;
+
+public class WatchlistItem
+{
+    public required string Group { get; set; }
+    public required string Name { get; set; }
+}
