+++ Financial.Application/Services/NavigationService.cs
@@ -7,7 +7,7 @@ using System.Linq;
 
 namespace Financial.Application.Services;
 
-public sealed class NavigationService : INavigationService, ICreditQueryService
+public sealed class NavigationService : INavigationService
 {
     private readonly IRepository _repository;
 
@@ -98,33 +98,4 @@ public sealed class NavigationService : INavigationService, ICreditQueryService
             .ToList();
     }
 
-    public IReadOnlyList<CreditDTO> GetCreditsByBroker(string brokerName)
-    {
-        if (string.IsNullOrWhiteSpace(brokerName))
-        {
-            return Array.Empty<CreditDTO>();
-        }
-
-        return _repository.GetAssetsByBroker(brokerName)
-            .Where(asset => asset.Active)
-            .SelectMany(asset => asset.Credits)
-            .Select(NavigationMapper.MapCredit)
-            .OrderByDescending(credit => credit.Date)
-            .ToList();
-    }
-
-    public IReadOnlyList<CreditDTO> GetCreditsByPortfolio(string brokerName, string portfolioName)
-    {
-        if (string.IsNullOrWhiteSpace(brokerName) || string.IsNullOrWhiteSpace(portfolioName))
-        {
-            return Array.Empty<CreditDTO>();
-        }
-
-        return _repository.GetAssetsByBrokerPortfolio(brokerName, portfolioName)
-            .Where(asset => asset.Active)
-            .SelectMany(asset => asset.Credits)
-            .Select(NavigationMapper.MapCredit)
-            .OrderByDescending(credit => credit.Date)
-            .ToList();
-    }
 }
