+++ Financial.Domain/Entities/Credit.cs
@@ -27,7 +27,7 @@ public class Credit
     public static Credit CreateWithId(Guid id, DateTime date, CreditType type, decimal value) =>
         new(id, date, type, value);
 
-    public void EnsureId()
+    internal void EnsureId()
     {
         if (Id == Guid.Empty)
         {
